Add unit/integration tests, validation to handler, and refactor MaxAsync method in handler/repository

+++ VictoryCenter/VictoryCenter.BLL/DTOs/TeamMembers/CreateTeamMemberDto.cs
@@ -10,7 +10,7 @@ public class CreateTeamMemberDto
 
     public string? MiddleName { get; set; }
     
-    public required int CategoryId { get; set; } 
+    public required long CategoryId { get; set; } 
     
     public Status Status { get; set; }
